explicit integer conversion.
use -Wconversion -Wno-sign-conversion to detect unwanted conversion.
or -Wshorten-64-to-32 in clang to avoid integer truncation.

+++ inc/ojlibs/graph.hpp
@@ -23,7 +23,7 @@ struct graph {
     std::vector<int> heads;
     std::vector<edge_type> edges;
 
-    int vertex_size() const { return heads.size(); }
+    int vertex_size() const { return (int)heads.size(); }
     int edge_size() const { return top; }
     void add_edge(int s, int t, const EInfo &info = EInfo()) {
         int edge_idx = top++;
